requirements-python-2.txt
future==0.15.2
PyYAML
Jinja2
